fixed workers can replace unbreakable blocks

+++ src/main/java/com/talhanation/workers/entities/ai/FarmerAI.java
@@ -9,12 +9,14 @@ import net.minecraft.tags.BlockTags;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.entity.ai.goal.Goal;
 import net.minecraft.world.item.BoneMealItem;
+import net.minecraft.world.item.HoeItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.FluidState;
 import net.minecraft.world.level.material.Fluids;
+import net.minecraftforge.common.Tags;
 import net.minecraftforge.event.ForgeEventFactory;
 
 import java.util.EnumSet;
@@ -215,28 +217,33 @@ public class FarmerAI extends Goal {
     }
     private void prepareWaterPos(){
         if (waterPos != null) {
+            BlockState blockState = this.farmer.getCommandSenderWorld().getBlockState(waterPos);
             FluidState waterBlockState = this.farmer.getCommandSenderWorld().getFluidState(waterPos);
 
             if (waterBlockState != Fluids.WATER.defaultFluidState() || waterBlockState != Fluids.FLOWING_WATER.defaultFluidState()){
-                farmer.getCommandSenderWorld().setBlock(waterPos, Blocks.WATER.defaultBlockState(), 3);
-                farmer.getCommandSenderWorld().playSound(null, waterPos.getX(), waterPos.getY(), waterPos.getZ(),
-                        SoundEvents.BUCKET_EMPTY, SoundSource.BLOCKS, 1.0F, 1.0F);
+                if(FarmerEntity.TILLABLES.contains(blockState.getBlock())){
+                    farmer.getCommandSenderWorld().setBlock(waterPos, Blocks.WATER.defaultBlockState(), 3);
+                    farmer.getCommandSenderWorld().playSound(null, waterPos.getX(), waterPos.getY(), waterPos.getZ(),
+                            SoundEvents.BUCKET_EMPTY, SoundSource.BLOCKS, 1.0F, 1.0F);
+                }
+                else return;
             }
             farmer.workerSwingArm();
         }
     }
     private void prepareFarmLand(BlockPos blockPos) {
         // Make sure the center block remains waterlogged.
-
-        if(blockPos != waterPos) {
+        BlockState blockState = this.farmer.getCommandSenderWorld().getBlockState(blockPos);
+        Block block = blockState.getBlock();
+        if(blockPos != waterPos && FarmerEntity.TILLABLES.contains(block)) {
             farmer.getCommandSenderWorld().setBlock(blockPos, Blocks.FARMLAND.defaultBlockState(), 3);
             farmer.getCommandSenderWorld().playSound(null, blockPos.getX(), blockPos.getY(), blockPos.getZ(), SoundEvents.HOE_TILL,
                     SoundSource.BLOCKS, 1.0F, 1.0F);
-            BlockState blockState = this.farmer.getCommandSenderWorld().getBlockState(blockPos.above());
-            Block block = blockState.getBlock();
+            BlockState blockStateAbove = this.farmer.getCommandSenderWorld().getBlockState(blockPos.above());
+            Block blockAbove = blockStateAbove.getBlock();
             farmer.workerSwingArm();
 
-            if (block instanceof BushBlock || block instanceof GrowingPlantBlock) {
+            if (blockAbove instanceof BushBlock || blockAbove instanceof GrowingPlantBlock) {
                 farmer.getCommandSenderWorld().destroyBlock(blockPos.above(), false);
                 farmer.getCommandSenderWorld().playSound(null, blockPos.getX(), blockPos.getY(), blockPos.getZ(), SoundEvents.GRASS_BREAK,
                         SoundSource.BLOCKS, 1.0F, 1.0F);
@@ -256,7 +263,7 @@ public class FarmerAI extends Goal {
 
                 Block block = blockState.getBlock();
 
-                boolean canSustainSeeds = block == Blocks.GRASS_BLOCK || block == Blocks.DIRT;
+                boolean canSustainSeeds = FarmerEntity.TILLABLES.contains(block);
                 boolean hasSpaceAbove = (aboveBlockState.is(Blocks.AIR) || aboveBlockState.is(BlockTags.REPLACEABLE));
 
                 if (canSustainSeeds && hasSpaceAbove) {
@@ -266,6 +273,8 @@ public class FarmerAI extends Goal {
         }
         return null;
     }
+
+
 
     public BlockPos getPlantPos() {
         // int range = 8;
